pyproject.toml
[project]
name = "climate-intelligence-copilot"
version = "0.1.0"
description = "AI-powered climate decision support system combining satellite data, LLMs, and agentic reasoning"
authors = [{name = "Climate Intelligence Team", email = "team@climate-copilot.ai"}]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "langchain>=0.1.0",
    "crewai>=0.0.1",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "rasterio>=1.3.0",
    "geopandas>=0.13.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.2.0",
    "jupyter>=1.0.0",
]

gee = [
    "ee>=0.2",
    "google-cloud-storage>=2.10.0",
]

satellite = [
    "sentinelsat>=1.3.0",
    "xarray>=2023.0.0",
    "netcdf4>=1.6.0",
]

[project.scripts]
climate-copilot = "main:cli"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
